// client/src/models/transactions.ts
export interface BankTransactions {
    dataVersion: number;
    reference: string;
    submissionTime: string;
    banks: Bank[];
    decisionMetrics: DecisionMetric[];
  }
  
  export interface Bank {
    bankName: string;
    bankSlug: string;
    errors: any[];
    bankAccounts: BankAccount[];
    address: { [key: string]: string };
  }
  
  export interface BankAccount {
    id: number;
    accountType: string;
    accountHolder: string;
    accountHolderType: string;
    accountName: string;
    bsb: string;
    accountNumber: string;
    currentBalance: string;
    availableBalance: string;
    transactions: Transaction[];
    statementSummary: StatementSummary;
    statementAnalysis: StatementAnalysis[];
    additionalDetails: AdditionalDetails;
  }
// ...
  export enum AnalysisPointType {
    Date = "date",
    Integer = "integer",
    Money = "money",
    String = "string",
  }
// ...
  export interface DecisionMetric {
    id: string;
    name: string;
    descriptor: string;
    type: AnalysisPointType;
    value: number | string;
  }